+++ packages/e2e-test-kit/package.json
@@ -25,7 +25,7 @@
     "react-dom": "^16.4.1",
     "rimraf": "^2.6.2",
     "@stylable/core": "^0.1.0",
-    "stylable-runtime": "^1.0.7",
+    "@stylable/runtime": "^0.1.0",
     "webpack": "^4.12.1",
     "webpack-sources": "^1.1.0"
   },
