+++ packages/runtime/src/cached-node-renderer.ts
