+++ packages/stylable-scripts/template/src/globals.d.ts
@@ -1,4 +1,4 @@
-type RuntimeStylesheet = import('stylable-runtime').RuntimeStylesheet;
+type RuntimeStylesheet = import('@stylable/runtime').RuntimeStylesheet;
 
 declare module '*.st.css' {
     const stylesheet: RuntimeStylesheet;
