+++ packages/stylable-webpack-plugin/package.json
@@ -14,7 +14,7 @@
     "css-selector-tokenizer": "^0.7.0",
     "find-config": "^1.0.0",
     "lodash.clonedeep": "^4.5.0",
-    "stylable-runtime": "^1.0.7",
+    "@stylable/runtime": "^0.1.0",
     "webpack-sources": "^1.1.0"
   },
   "devDependencies": {
